Explicitly link against jbig2dec, openjp2, and jpeg
Normally these are statically linked into libmupdfthird, but we delete
those to make mupdf use the system libraries.  Previously
zathura-pdf-mupdf explicitly linked against them, but this is
technically incorrect since they are supposed to be in libmupdfthird so
that was removed.  This commit essentially reverts that in the new build
system.